+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-		Markdown Markup Parser (MDMP) - v0.6.1
+		Markdown Markup Parser (MDMP) - v0.8.0
 		Created by: Trevor W.
 
 		Github: https://github.com/trevor34/markdown-markup-parser/
@@ -17,7 +17,10 @@ All notable changes to this project will be documented in this file. The format 
 
 
 ## [Unreleased] - Upcoming Changes, Current Projects, and 'wish list' items
-- Add commands for a CSS framework
+- Add HTML tag support besides what is parseable with markdown-it
+## [0.8] 2018-7-16
+## Added
+- Added support for selectors on items besides divs with the `/!selector` command
 ## [0.7] 2018-7-07
 ## Changed
 - Changed GitHub title to Markdown Markup Parser from Markdown Website Builder